+++ shared/gpu_kernels.cuh
@@ -175,6 +175,6 @@ __global__ void mixLabels(unsigned int * activeNodes, bool *label1, bool *label2
 
 __global__ void moveUpLabels(unsigned int * activeNodes, bool *label1, bool *label2, unsigned int size, unsigned int from);
 
-__global__ void moveWalkers_pr(unsigned int num_nodes, int *numWalker1, int *numWalker2, float *value);
+__global__ void moveWalkers_pr(unsigned int num_nodes, int *numWalker1, int *numWalker2, u_int32_t *value);
 
 __global__ void moveWalkers_ppr(unsigned int num_nodes, int *numWalker1, int *numWalker2, float *value, curandState *randStates);
